+++ src/hooks/PlayLayer.cpp
@@ -48,9 +48,10 @@ void HookPlayLayer::updateStartPos(int idx) {
 void HookPlayLayer::createObjectsFromSetupFinished() {
     PlayLayer::createObjectsFromSetupFinished();
 
-    if(this->m_startPosObject) {
-        auto currentIdx = find(m_fields->m_startPosObjects.begin(), m_fields->m_startPosObjects.end(), this->m_startPosObject) - m_fields->m_startPosObjects.begin();
-        m_fields->m_startPosIdx = currentIdx + 1;
+    if(m_startPosObject) {
+        auto currentIdx = std::find(m_fields->m_startPosObjects.begin(), m_fields->m_startPosObjects.end(), this->m_startPosObject);
+        if(currentIdx != m_fields->m_startPosObjects.end())
+         m_fields->m_startPosIdx = std::distance(m_fields->m_startPosObjects.begin(), currentIdx) - 1;
     }
 
     std::sort(m_fields->m_startPosObjects.begin(), m_fields->m_startPosObjects.end(), [](auto* a, auto* b) { return a->getPositionX() < b->getPositionX(); });