+++ changelog.md
@@ -1,4 +1,7 @@
 # Startpos Switcher Changelog
+## v1.0.4
+- Fixed incorrect current Startpos calculation
+- Added custom keybinds
 ## v1.0.3
 - Added don't fade on start
 - Added hide buttons